servico_financeiro/requirements.txt
asgiref==3.3.1
Django==3.1.6
djangorestframework==3.12.2
pika==1.2.0
psycopg2-binary==2.8.5
python-decouple==3.3
pytz==2021.1
sqlparse==0.4.1